change type

+++ sacabench/util/bits.hpp
@@ -61,7 +61,7 @@ inline uint64_t powi(uint64_t base, uint64_t exp) {
 inline uint64_t floor_log2(uint64_t x) { return floor(std::log2(x)); }
 
 /// Calculates next power of 2.
-inline size_t next_pow2(size_t x) {
+inline uint64_t next_pow2(uint64_t x) {
     if (x < 2)
         return 1;
     return 1ull << (64 - __builtin_clzll(x - 1));
